+++ src/libs/protocols/index.ts
@@ -6,6 +6,7 @@ import { KriyaContract } from "./kriya";
 import { TurbosContract } from "./turbos";
 import { SuiswapContract } from "./suiswap";
 import { BluemoveContract } from "./bluemove";
+import { KriyaV3Contract } from "./kriyaV3";
 
 export const ProtocolContract = {
   cetus: CetusContract,
@@ -16,4 +17,5 @@ export const ProtocolContract = {
   aftermath: AfterMathContract,
   deepbook: DeepBookContract,
   flowx: FlowXContract,
+  kriya_v3: KriyaV3Contract,
 };